fix: update duplicate vote handling
Instead of disacrding ALL votes for an address that has duplicates, first check if they are all to the same address, and in that case count as one valid vote and mark the remaining invalid.

If an address votes for both yes and no, then the vote will be discarded.